+++ App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Navigator from './navigation';
+import Navigator from './config/navigation';
 
 const Main = () => <Navigator />;
